ticket:11018: Support CLDR time separator (ICU4J)
This change will treat the ':' character in patterns as a symbol
(instead of as a literal) and if the Locale defines timeSeparator then
use that character instead when formatting and accept both when parsing.

It will also define the '.' as an alternate time separator, that will be
accepted when parsing in lenient mode.

R=rocketman@google.com

Review URL: https://codereview.appspot.com/180370043
X-Commit-URL: https://ssl.icu-project.org/trac/changeset/36898
Former-commit-id: 53a50908bf035bafb98cfa098f2886a7f67163c4